完成了三角波和锯齿波的函数测试，可以输出每20次的步径值，还没有在电压板测试

+++ Inc/dac7821.h
@@ -11,7 +11,7 @@
 void DAC7821_init(void);
 void DAC7821_vol(float vol);
 void data_send_DAC7821(uint16_t data);
-void sanjiaobo(float fre,float amp1,float amp2);
-void juchibo(float fre,float amp1,float amp2);
+void sanjiaobo(uint16_t fre,uint8_t amp1,uint8_t amp2);
+void juchibo(uint16_t fre,uint8_t amp1,uint8_t amp2);
 void DAC7821_ShutDown();
 #endif
